Add files via upload
New-WPFMessageBox.ps1 is the edited PowerShell script that handles the WPF calls. New-WPFMessageBox.ps1.bin.bas is the embedded binary BAS file that is added in your '$INCLUDE at the beginning of your programs. WPFMessageBox.bas contains the function I created as well as a small console test program for you to try it out and see how it runs with an _ECHO back to the console so you can see how it is calling the script.