Edited BM SHELL$ and added TYPEs to BI
Added custom TYPE variables that can be accessed as dotted variables in QB64. Ex: Hue.White would give the color "White" for the message box. This is to reduce typos that could lead to no message box being generated. Also, parameters that utilize numerical values are now stored as integers so as to prevent accidentally typing a letter instead of a number for those.