Fixed issue with desktop shortcut
Desktop shortcuts were causing the message boxes to not appear in programs because the _STARTDIR$ value was pointing to the desktop rather than the start directory of the EXE. I have changed the way that it passes the start directory so now it should work fine. Also, I have moved the Replace functions from the PS1 file to the BM. Now I pass the string with the HTML hex code for the new line and tab characters before the script is ran.